+++ src/parser/generate_tree.rs
@@ -67,7 +67,7 @@ pub fn generate_tree(mut lexer: Lexer, analysis_table: &AnalysisTable) -> (Rc<Re
                                 input.line,
                                 input.start_char,
                                 if input.len > 0 {
-                                    input.start_char + (input.len - 1) as u64
+                                    input.start_char + input.len - 1
                                 } else {
                                     input.start_char
                                 },
@@ -132,7 +132,7 @@ pub fn generate_tree(mut lexer: Lexer, analysis_table: &AnalysisTable) -> (Rc<Re
                                     input.line,
                                     input.start_char,
                                     if input.len > 0 {
-                                        input.start_char + (input.len - 1) as u64
+                                        input.start_char + input.len - 1
                                     } else {
                                         input.start_char
                                     },
@@ -160,7 +160,7 @@ pub fn generate_tree(mut lexer: Lexer, analysis_table: &AnalysisTable) -> (Rc<Re
                         input.line,
                         input.line,
                         input.start_char,
-                        input.start_char + (input.len - 1) as u64,
+                        input.start_char + input.len - 1,
                         "Stack is empty and input is not EOF".to_string(),
                     )
                     .display();
